requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
brotli=1.1.0=hb547adb_1
brotli-bin=1.1.0=hb547adb_1
brotli-python=1.1.0=py312h9f69965_1
bzip2=1.0.8=h93a5062_5
ca-certificates=2024.7.4=hf0a4a13_0
certifi=2024.7.4=pyhd8ed1ab_0
cffi=1.16.0=py312h8e38eb3_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
contourpy=1.2.1=py312h0fef576_0
cycler=0.12.1=pyhd8ed1ab_0
filelock=3.15.4=pyhd8ed1ab_0
fonttools=4.53.1=py312h7e5086c_0
freetype=2.12.1=hadb7bae_2
fsspec=2024.6.1=pyhff2d567_0
gmp=6.3.0=h7bae524_2
gmpy2=2.1.5=py312hfa9fade_1
h2=4.1.0=pyhd8ed1ab_0
hpack=4.0.0=pyh9f0ad1d_0
hyperframe=6.0.1=pyhd8ed1ab_0
idna=3.7=pyhd8ed1ab_0
jinja2=3.1.4=pyhd8ed1ab_0
kiwisolver=1.4.5=py312h389731b_1
lcms2=2.16=ha0e7c42_0
lerc=4.0.0=h9a09cb3_0
libabseil=20240116.2=cxx17_hebf3989_0
libblas=3.9.0=22_osxarm64_openblas
libbrotlicommon=1.1.0=hb547adb_1
libbrotlidec=1.1.0=hb547adb_1
libbrotlienc=1.1.0=hb547adb_1
libcblas=3.9.0=22_osxarm64_openblas
libcxx=17.0.6=h0812c0d_3
libdeflate=1.20=h93a5062_0
libexpat=2.6.2=hebf3989_0
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=13_2_0_hd922786_3
libgfortran5=13.2.0=hf226fd6_3
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=22_osxarm64_openblas
libopenblas=0.3.27=openmp_h517c56d_1
libpng=1.6.43=h091b4b1_0
libprotobuf=4.25.3=hbfab5d5_0
libsqlite=3.46.0=hfb93653_0
libtiff=4.6.0=h07db509_3
libtorch=2.3.1=cpu_generic_h95df8ed_1
libuv=1.48.0=h93a5062_0
libwebp-base=1.4.0=h93a5062_0
libxcb=1.16=hf2054a2_0
libzlib=1.3.1=hfb2fe0b_1
llvm-openmp=18.1.8=hde57baf_0
markupsafe=2.1.5=py312he37b823_0
matplotlib=3.8.4=py312h1f38498_2
matplotlib-base=3.8.4=py312h4479663_2
mpc=1.3.1=h91ba8db_0
mpfr=4.2.1=h41d338b_1
mpmath=1.3.0=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.5=hb89a1cb_0
networkx=3.3=pyhd8ed1ab_1
nomkl=1.0=h5ca1d4c_0
numpy=2.0.0=py312hb544834_0
openjpeg=2.5.2=h9f1df11_0
openssl=3.3.1=hfb2fe0b_1
packaging=24.1=pyhd8ed1ab_0
pillow=10.4.0=py312h39b1d8d_0
pip=24.0=pyhd8ed1ab_0
pthread-stubs=0.4=h27ca646_1001
pycparser=2.22=pyhd8ed1ab_0
pyparsing=3.1.2=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.12.4=h30c5eda_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python_abi=3.12=4_cp312
pytorch=2.3.1=cpu_generic_py312h2aa0b4f_1
readline=8.2=h92ec313_1
requests=2.32.3=pyhd8ed1ab_0
scipy=1.14.0=py312h14ffa8f_1
setuptools=70.1.1=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sleef=3.5.1=h156473d_2
sympy=1.12.1=pypyh2585a3b_103
tk=8.6.13=h5083fa2_1
torchvision=0.18.1=cpu_py312h5f70645_1
tornado=6.4.1=py312h7e5086c_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024a=h0c530f3_0
urllib3=2.2.2=pyhd8ed1ab_1
wheel=0.43.0=pyhd8ed1ab_1
xorg-libxau=1.0.11=hb547adb_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.6=h57fd34a_0
zstandard=0.22.0=py312h721a963_1
zstd=1.5.6=hb46c0d2_0
